refactor: expenses, previous-challenges 페이지 home 하위로 이동
