+++ src/app/home/expenses/add/_components/AmountInput.type.ts
